+++ src/main/java/com/project/Wayg/WaygApplication.java
@@ -2,8 +2,6 @@ package com.project.Wayg;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 @SpringBootApplication
 public class WaygApplication{
