+++ src/main/java/com/project/Wayg/repository/WaygRepository.java
@@ -1,13 +1,9 @@
-package com.project.Wayg.Repository;
+package com.project.Wayg.repository;
 
 import com.project.Wayg.Entity.School;
-import com.project.Wayg.Entity.dto.RequestDTO;
-import com.project.Wayg.config.QuerydslConfiguration;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.querydsl.QuerydslPredicateExecutor;
-import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
