+++ src/main/resources/application.properties
@@ -14,3 +14,5 @@ spring.jpa.open-in-view=false
 spring.jpa.properties.hibernate.show_sql=true
 spring.jpa.properties.hibernate.format_sql=true
 
+
+
